Moved app.js and app.css into src-frontend and updated reference in index.js
