+++ src/src-frontend/Components/Searchbar/Searchbar.js
@@ -1,6 +1,5 @@
 import React from "react";
 import "./Searchbar.css";
-import { Router, Route, Switch } from "react-router-dom";
 
 export default function Searchbar() {
   return <div>Hello, searchbar!</div>;
